+++ test/core_tests/group_address_dpt_test.py
@@ -30,7 +30,7 @@ async def test_group_address_dpt_in_telegram_queue(xknx_no_interface):
         assert telegrams == 1
         assert test_telegram.decoded_data is None
 
-        xknx.group_address_dpt.set_dpts({GroupAddress("1/2/3"): {"main": 5, "sub": 1}})
+        xknx.group_address_dpt.set({GroupAddress("1/2/3"): {"main": 5, "sub": 1}})
         await xknx.telegrams.put(telegram)
         await xknx.telegrams.join()
         assert telegrams == 2
@@ -39,10 +39,37 @@ async def test_group_address_dpt_in_telegram_queue(xknx_no_interface):
         assert test_telegram.decoded_data.value == 50
 
 
-def test_set_dpts(xknx_no_interface):
-    """Test set_dpts."""
+@patch("logging.Logger.warning")
+async def test_get_invalid_payload(logger_warning_mock, xknx_no_interface):
+    """Test processing when DPT doesn't fit payload."""
+
+    test_telegram: Telegram = None
+
+    async def _telegram_callback(telegram: Telegram):
+        nonlocal test_telegram
+        test_telegram = telegram
+
     xknx = xknx_no_interface
-    xknx.group_address_dpt.set_dpts(
+    xknx.group_address_dpt.set({GroupAddress("1/2/3"): {"main": 5}})
+    xknx.telegram_queue.register_telegram_received_cb(_telegram_callback)
+
+    telegram = Telegram(
+        destination_address=GroupAddress("1/2/3"),
+        direction=TelegramDirection.INCOMING,
+        payload=apci.GroupValueWrite(DPTArray((0x01, 0x02))),  # wrong payload size
+    )
+    async with xknx:
+        await xknx.telegrams.put(telegram)
+        await xknx.telegrams.join()
+        assert test_telegram.decoded_data is None
+        assert logger_warning_mock.call_count == 1
+        assert "DPT decoding error" in logger_warning_mock.call_args[0][0]
+
+
+def test_set(xknx_no_interface):
+    """Test set."""
+    xknx = xknx_no_interface
+    xknx.group_address_dpt.set(
         {
             GroupAddress("1/2/3"): {"main": 5, "sub": 1},
             1: "temperature",
@@ -54,17 +81,15 @@ def test_set_dpts(xknx_no_interface):
         1: DPTTemperature,
         "i-internal": DPTHumidity,
     }
-    assert (
-        xknx.group_address_dpt.get_transcoder(GroupAddress("0/0/1")) is DPTTemperature
-    )
+    assert xknx.group_address_dpt.get(GroupAddress("0/0/1")) is DPTTemperature
 
 
 @patch("logging.Logger.warning")
 @patch("logging.Logger.debug")
-def test_set_dpts_invalid(logger_debug_mock, logger_warning_mock, xknx_no_interface):
+def test_set_invalid(logger_debug_mock, logger_warning_mock, xknx_no_interface):
     """Test set invalid dpts."""
     xknx = xknx_no_interface
-    xknx.group_address_dpt.set_dpts(
+    xknx.group_address_dpt.set(
         {
             GroupAddress("0/0/1"): {"main": None},
             2: "invalid",
@@ -76,3 +101,17 @@ def test_set_dpts_invalid(logger_debug_mock, logger_warning_mock, xknx_no_interf
     assert logger_debug_mock.call_count == 1
     assert "No transcoder found for DPTs" in logger_debug_mock.call_args[0][0]
     assert not xknx.group_address_dpt._ga_dpts
+
+
+def test_clear(xknx_no_interface):
+    """Test clear."""
+    xknx = xknx_no_interface
+    xknx.group_address_dpt.set(
+        {
+            GroupAddress("1/2/3"): {"main": 5, "sub": 1},
+            1: "temperature",
+        }
+    )
+    assert len(xknx.group_address_dpt._ga_dpts) == 2
+    xknx.group_address_dpt.clear()
+    assert len(xknx.group_address_dpt._ga_dpts) == 0